agent-createcharacter/requirements.txt
fastapi
uvicorn[standard]
google-cloud-firestore
python-dotenv
httpx
google-adk>=1.27.0
google-genai>=1.9.0
a2a
fastmcp
pydantic
